docker_compose/movies_admin/requirements.txt
django>=4.2.11, <5.0.0
django-debug-toolbar==4.3.0
django-split-settings==1.3.0
Faker==24.3.0
psycopg2-binary==2.9
python-dotenv==1.0.1
uWSGI==2.0.24
